Fix incompatible int type

+++ JSON/json/json.h
@@ -825,7 +825,7 @@ private:
 #endif
   } value_;
   ValueType type_ : 8;
-  uint allocated_ : 1; // Notes: if declared as bool, bitfield is useless.
+  uint8_t allocated_ : 1; // Notes: if declared as bool, bitfield is useless.
 #ifdef JSON_VALUE_USE_INTERNAL_MAP
   unsigned int itemIsUsed_ : 1; // used by the ValueInternalMap container.
   int memberNameIsStatic_ : 1;  // used by the ValueInternalMap container.
